+++ packages/components/hooks/useOutsideClick.ts
@@ -4,18 +4,22 @@ const useOutsideClick = (callback: () => void) => {
   const ref = React.useRef<HTMLDivElement>(null);
 
   React.useEffect(() => {
-    const handleClick = (event: MouseEvent) => {
+    const handleClickOutside = (event: MouseEvent | TouchEvent) => {
       if (ref.current && !ref.current.contains(event.target as Node)) {
         callback();
       }
     };
 
-    document.addEventListener("click", handleClick, true);
+    // document.addEventListener("click", handleClick, true);
+    document.addEventListener("mousedown", handleClickOutside, true);
+    document.addEventListener("touchstart", handleClickOutside, true); // Ensure touch events are handled
 
     return () => {
-      document.removeEventListener("click", handleClick, true);
+      // document.removeEventListener("click", handleClick, true);
+      document.removeEventListener("mousedown", handleClickOutside, true);
+      document.removeEventListener("touchstart", handleClickOutside, true);
     };
-  }, [ref]);
+  }, [ref, callback]);
 
   return ref;
 };